backend/requirements.txt
annotated-types==0.7.0
anyio==4.3.0
asgiref==3.8.1
attrs==23.2.0
Automat==22.10.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
constantly==23.10.4
cryptography==42.0.7
cssselect==1.2.0
defusedxml==0.7.1
distro==1.9.0
dnspython==2.6.1
email_validator==2.1.1
fake-useragent==1.5.1
Faker==25.2.0
fastapi==0.111.0
fastapi-cli==0.0.4
filelock==3.14.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
itemadapter==0.9.0
itemloaders==1.2.0
Jinja2==3.1.4
jmespath==1.0.1
lxml==5.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
openai==1.30.1
orjson==3.10.3
packaging==24.0
parsel==1.9.1
Protego==0.3.1
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
PyDispatcher==2.0.7
Pygments==2.18.0
pyOpenSSL==24.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
queuelib==1.7.0
requests==2.31.0
requests-file==2.0.0
rich==13.7.1
Scrapy==2.11.2
scrapy-fake-useragent==1.4.4
service-identity==24.1.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
starlette==0.37.2
tldextract==5.1.2
tqdm==4.66.4
Twisted==24.3.0
typer==0.12.3
typing_extensions==4.11.0
ujson==5.10.0
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
w3lib==2.1.2
watchfiles==0.21.0
websockets==12.0
zope.interface==6.4
